// src/enums/AxisUnit.ts
/**
 * Axis unit enumeration for positioning calculations
 */
export enum AxisUnit {
  PIXEL = 'pixel',
  PERCENTAGE = 'percentage',
  VIEWPORT = 'viewport',
  PARENT = 'parent',
  AUTO = 'auto',
  CENTER = 'center',
  LEFT = 'left',
  RIGHT = 'right',
  TOP = 'top',
  BOTTOM = 'bottom',
  X = 'x',
  Y = 'y'
}